refactor(ims): extract the inbound-call lifecycle model and drive it live
Move CallStage/EndedBy/BridgedCall/Admission/Maintenance*/ServiceHealth out of
volte::bridge — where they were exercised only by tests — into a
transport-agnostic ims::lifecycle, so one model serves both the Wi-Fi and
cellular paths through the shared ims::agent dispatch loop.

Wire the model into the live path:
- busy admission uses the shared Admission::for_current rule
- ActiveCall carries a BridgedCall; every teardown attributes the end cause via
  EndedBy and derives the recorded outcome/CallStatus from the model
- MaintenancePolicy replaces the ad-hoc active_call.is_some() renewal-deferral
  gate and exposes deferral for status
- ServiceHealth (via RegistrationStatus::health()) backs volte-status's new
  can_answer/blocked_reason, wired over backward-compatible #[serde(default)]
  RegistrationStatusReply fields

Reporting is derived from the model rather than restated at each teardown site,
without collapsing the fine-grained BridgeFailureReason into the coarser
EndedBy (that would lose ring-timeout-vs-declined fidelity); EndedBy::as_str and
the reason:: control strings are asserted to coincide where they overlap.

Also split PANI out of volte::mod into volte::pani, and DRY the six +CMD
AT-line parsers behind pdn::at_fields.

Validated live on a bridged VoLTE call (ended_by=pbx_hangup, outcome=answered)
and the pre-answer decline path; make lint clean, workspace tests green.
